+++ src/main/kotlin/io/github/dylmeadows/eontimer/model/timer/Gen3TimerModel.kt
@@ -2,24 +2,21 @@ package io.github.dylmeadows.eontimer.model.timer
 
 import io.github.dylmeadows.eontimer.util.getValue
 import io.github.dylmeadows.eontimer.util.setValue
-import javafx.beans.property.IntegerProperty
-import javafx.beans.property.ObjectProperty
-import javafx.beans.property.SimpleIntegerProperty
-import javafx.beans.property.SimpleObjectProperty
+import javafx.beans.property.*
 
 class Gen3TimerModel {
     val modeProperty: ObjectProperty<Gen3TimerMode> = SimpleObjectProperty(Gen3TimerConstants.DEFAULT_MODE)
     var mode: Gen3TimerMode by modeProperty
 
-    val calibrationProperty: IntegerProperty = SimpleIntegerProperty(Gen3TimerConstants.DEFAULT_CALIBRATION)
+    val calibrationProperty: LongProperty = SimpleLongProperty(Gen3TimerConstants.DEFAULT_CALIBRATION)
     var calibration by calibrationProperty
 
-    val preTimerProperty: IntegerProperty = SimpleIntegerProperty(Gen3TimerConstants.DEFAULT_PRE_TIMER)
+    val preTimerProperty: LongProperty = SimpleLongProperty(Gen3TimerConstants.DEFAULT_PRE_TIMER)
     var preTimer by preTimerProperty
 
-    val targetFrameProperty: IntegerProperty = SimpleIntegerProperty(Gen3TimerConstants.DEFAULT_TARGET_FRAME)
+    val targetFrameProperty: LongProperty = SimpleLongProperty(Gen3TimerConstants.DEFAULT_TARGET_FRAME)
     var targetFrame by targetFrameProperty
 
-    val frameHitProperty: IntegerProperty = SimpleIntegerProperty()
+    val frameHitProperty: LongProperty = SimpleLongProperty()
     var frameHit by frameHitProperty
 }
